requirements.txt
anthropic>=0.80.0
chromadb==0.5.23
fastapi==0.115.4
uvicorn==0.32.0
google-genai>=1.0.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==5.1.0
